Main

Description

Alumno: Cazajous Miguel A.

Asignatura: Paradigmas de programación

Docente: Ing. Wolfmann Gustavo

Tema:

Algoritmo minimax para elección de mejor movimiento en el juego TA TE TI

Requisitos:

Se necesita instalar ghc y el módulo random.

sudo pacman -S ghc (Para distribuciones basadas en Arch)

Para el módulo descargarlo de https://hackage.haskell.org/package/random

y seguir las instrucciones de https://wiki.haskell.org/Cabal/How_to_install_a_Cabal_package

Descripción:

El programa comienza tomando un archivo del cual lee un estado de juego

muestra el contenido obtenido representado en un tablero y verifica que este cumpla con un estado válido,

de no serlo el programa finaliza, de ser válido continua y verifica si hay un ganador (de haberlo lo muestra por pantalla)

(es posible que el estado sea de un juego finalizado) si no hay sigue y comprueba que jugador debe continuar

(el que tiene un movimiento menos) el jugador en caso de que tengan la misma cantidad de movimientos es el jugador O

el cual es el jugador "imposible".

El jugador X elige entre los posibles valores disponibles uno al azar.

De acuerdo a como está implementado es posible que el jugador O pierda si el siguiente turno es del jugador X

y la función random justo elige el valor que lo consagra ganador, de otra forma el jugador O que es el que ejecuta el algoritmo

Minimax será el ganador.

Para algunos estados de juego donde el jugador O puede ganar con un simple movimiento se puede observar que esto no ocurre siempre

y se debe a que cuando se consideran los puntajes solo se verifica que X no sea el ganador de otra forma le asigna el mismo puntaje

a un movimiento que lo consagre ganador instantáneamente y a otro que no y luego se elige el primero de la lista.

Por ese motivo puede demorarse más de la lógica la victoria del jugador O

Para ejecutar el programa se hizo uso de las tuberías.

cat archivo | runhaskell programa.hs

Donde "archivo" representa a los diferentes estados del juego

Documentation

main :: IO ()

Función principal. Lee un archivo con un estado del juego,

Invoca a la función unaLinea y luego a la función juego

unaLinea :: String -> String

Omite los espacios y saltos de línea que se leyeron del archivo

Recibe un String (xs) y remueve cada caracter que sea igual a un salto de línea

o un espacio.

juego :: String -> IO ()

Llama a mostrarString para mostrar una representación del estado del juego y

luego verifica mediante formato que el estado recibido es correcto.

Verifica si hay un ganador llamando a la función ganador. Si el valor devuelto no es un caracter

espacio, es decir que hay un ganador, lo muestra. En caso contrario continua.

Si no se ha encontrado un ganador define cual es el jugador que debe continuar llamando a proxJugador

Para el caso del jugador X elige una posición aleatoria entre las que son válidas (movPermitidos)

y realiza el movimiento llamando a la función mover. Para el valor aleatorio tomamos la longitud total

de la lista devuelta por movPermitidos menos 1 (la lista comienza desde 0)

En el caso del jugador O la posición se determina llamando a mejorMovimiento

Luego de realizado el movimiento se juego se llama a si misma con el tablero ya modificado con el último movimiento

Para el caso en que el tablero esté vacío para acelerar ese primer movimiento simplemente se mueve el jugador O

(Ya que en caso de que esté vacío es el jugador por defecto) a la posición 4 del tablero (el centro)

mostrarString :: String -> String

Muestra una representación del tablero considerando el estado que se leyó desde archivo

formato :: String -> Bool

Verifica que el formato del estado de juego obtenido del archivo sea válido

Esto implica que no contenga elementos diferentes a X, O y E y además que

la diferencia entre cantidad de X y O no sea mayor a 1

Utiliza las funciones cantX, cantO y cantE

cantX :: String -> Int

Cuenta la cantidad de X que tiene el estado de juego

cantO :: String -> Int

Cuenta la cantidad de O que tiene el estado de juego

cantE :: String -> Int

Cuenta la cantidad de E que tiene el estado de juego

ganador :: String -> Char

Verifica cada línea vertical, horizontal y diagonal en busca de un ganador.

Si lo encuentra devuelve el ganador (X u O), sino devuelve caracter vacío.

Primero verifica que el comienzo de cada fila no sea una E (empty) y luego que los valores

de esa línea sean iguales. Si lo son devuelve ese valor que puede ser X o O.

De igual manera se procede con las columnas y con las diagonales.

proxJugador :: String -> Char

Define que jugador tiene el siguiente turno usando cantX y cantO

definiendo como siguiente jugador el que tenga un movimiento menos

movPermitidos :: String -> [Int]

Arma una lista con todas posiciones vacías.

Utiliza la función ganador y la función esValido

Recibe una cadena con un estado de juego, si existe un ganador devuelve una lista vacía

que representa que no hay movimientos válidos.

De lo contrario usando la función esValido devuelve una lista de enteros en las

posiciones donde todavía puede jugarse.

esValido :: String -> Int -> Bool

Retorna verdadero si el movimiento es válido

Recibe un entero que verifica que esté entre 0 y 9, y un estado de juego.

Si el valor coincide donde se encuentra el caracter E devuelve verdadero, de lo contrario

devuelve falso.

mover :: String -> Char -> Int -> String

Esta función realiza el movimiento del Char a la posición Int y devuelve un nuevo tablero.

mejorMovimiento :: String -> Int

Dado un tablero retorna una posición que asegura no se pierde

Utiliza la función puntajeMovimientos y luego invoca a maximoPuntaje

para obtener la posición deseada

puntajeMovimientos :: String -> [(Int, Int)]

Retorna una lista de tuplas (movimiento, puntaje)

Invoca a movPermitidos y a mayorPuntaje

mayorPuntaje :: String -> Int

Retorna el mayor puntaje de movimiento para el tablero recibido

Invoca a movPermitidos, puntajeString, mover y menorPuntaje

menorPuntaje :: String -> Int

Retorna el menor puntaje de movimiento para el tablero recibido

Invoca a movPermitidos, puntajeString, mover y mayorPuntaje

puntajeString :: String -> Char -> Int

Asigna un puntaje a un jugador recibido (Char).

maximoPuntaje :: (Int, Int) -> (Int, Int) -> (Int, Int)

De dos tuplas (movimiento, puntaje) devuelve la de mayor puntaje