Main

Description

Alumno: Cazajous Miguel A.

Asignatura: Paradigmas de programación

Docente: Ing. Wolfmann Gustavo

Tema:

Algoritmo minimax para elección de mejor movimiento en el juego TA TE TI

Documentation

type Tablero = String

Sinónimo de tipo

main :: IO ()

Función principal. Lee un archivo con un estado del juego,

llama a la función unaLinea y luego a la función juego

unaLinea :: Tablero -> String

Omite los espacios y saltos de línea que se leyeron del archivo

Y devuelve un String con 9 valores entre X, O y E

juego :: String -> IO ()

Llama a mostrarTablero y empieza el juego con un estado recibido desde main

.........

mostrarTablero :: Tablero -> String

mostrarTablero: Muestra una representación del tablero considerando el estado que se leyó desde archivo

scoreBoard :: Tablero -> Char -> Int

Asigna un puntaje

evaluateBoardMin :: Tablero -> Int

scores the board and returns minimum value move for the given board

evaluateBoardMax :: Tablero -> Int

scores the board and returns maximum value move for the given board

scoreMoves :: Tablero -> [(Int, Int)]

Retorna una lista de tuplas (movimiento, puntaje)

maxScore :: (Int, Int) -> (Int, Int) -> (Int, Int)

De dos tuplas (movimiento, puntaje) devuelve la de mayor puntaje

mejorMovimiento :: Tablero -> Int

Dado un tablero elige una posición que asegura no se pierde

esValido :: Tablero -> Int -> Bool

Retorna verdadero si el movimiento es válido

movPermitidos :: Tablero -> [Int]

Arma una lista con todas posiciones vacías.

ganador :: Tablero -> Char

Verifica cada línea vertical, horizontal y diagonal en busca de un ganador. Si lo encuentra devuelve el ganador (X u O), sino devuelve caracter vacío

formato :: Tablero -> Bool

Verifica que el formato del estado de juego obtenido del archivo sea válido

prox_jugador :: Tablero -> Char

Define que jugador tiene el siguiente turno

cantX :: Tablero -> Int

Cuenta la cantidad de X

cantO :: Tablero -> Int

Cuenta la cantidad de O

cantE :: Tablero -> Int

Cuenta la cantidad de E

mover :: Tablero -> Char -> Int -> Tablero

new mover applied