Main

Description

Alumno: Cazajous Miguel A.

Asignatura: Paradigmas de programación

Docente: Ing. Wolfmann Gustavo

Tema:

Algoritmo minimax para elección de mejor movimiento en el juego TA TE TI

Requisitos:

Se necesita instalar ghc y el módulo random.

sudo pacman -S ghc ghc-static (Para distribuciones basadas en Arch)

El módulo debe ser descargado de https://hackage.haskell.org/package/random

y se instala siguiendo las instrucciones de https://wiki.haskell.org/Cabal/How_to_install_a_Cabal_package

Ejecución:

Para ejecutar el programa usar make arg=file

Descripción:

El programa comienza tomando un archivo del cual lee un estado de juego, luego muestra el contenido

obtenido representado en un tablero y verifica que este cumpla con un estado válido, de no serlo el programa

finaliza con un mensaje de error, de lo contrario continua y verifica si hay un ganador (de haberlo lo muestra

por pantalla), si no hay, sigue y comprueba que jugador debe continuar jugando, dándole prioridad al jugador

O en caso de que exista la misma cantidad de O que se X.

El jugador X elige entre los posibles valores disponibles uno al azar haciendo uso del módulo Random.

Aunque el jugador O sea un jugador perfecto, pues es el que implementa el algoritmo minimax, aún es posible

que este pierda si el siguiente turno es del jugador X y la función random justo elige el valor

que lo consagra ganador, de otra forma el jugador O será el ganador.

Para algunos estados de juego donde el jugador O puede ganar con un simple movimiento se puede observar

que esto no ocurre siempre y se debe a que cuando se consideran los puntajes solo se verifica que X no

sea el ganador, pero no se considera la profundidad, siempre elige el primero valor de la lista de puntajes,

sea o no la forma más rápida de consagrarse ganador. Por ese motivo puede demorarse más la victoria del

jugador O.

Documentation

main :: IO ()

Función principal. Lee un archivo con un estado del juego,

Invoca a la función unaLinea y luego a la función juego

unaLinea :: String -> String

Omite los espacios y saltos de línea que se leyeron del archivo

Recibe un String (xs) y remueve cada caracter que sea igual a un salto de línea

o un espacio.

juego :: String -> IO ()

Llama a mostrarString para mostrar una representación del estado del juego y

luego verifica mediante formato que el estado recibido es correcto. Sino lo es sale del programa

con un mensaje de error.

Verifica si hay un ganador llamando a la función ganador. Si el valor devuelto no es un caracter

espacio, es decir que hay un ganador, lo muestra. En caso contrario continua.

Si no se ha encontrado un ganador define cual es el jugador que debe continuar llamando a proxJugador

Para el caso del jugador X elige una posición aleatoria entre las que son válidas (movPermitidos)

y realiza el movimiento llamando a la función mover. Para el valor aleatorio tomamos la longitud total

de la lista devuelta por movPermitidos menos 1 (la lista comienza desde 0). Finalmente vuelve a llamar

la función juego de manera recursiva para continuar con el juego.

Si el próximo jugador es O primero verifica si es el primer movimiento del juego, de serlo coloca O

en el centro del tablero, en caso contrario llama a la función mejorMovimiento.

Luego de realizado el movimiento se llama a juego nuevamente para continuar con el proceso.

mostrarString :: String -> String

Muestra una representación del tablero considerando el estado que se leyó desde el archivo

formato :: String -> Bool

Verifica que el formato del estado de juego obtenido del archivo sea válido

Esto implica que no contenga elementos diferentes a X, O y E y además que

la diferencia entre cantidad de X y O no sea mayor a 1

Utiliza las funciones cantX, cantO y cantE

cantX :: String -> Int

Cuenta la cantidad de X que tiene el estado de juego.

Suma 1 cada vez que un caracter de la cadena es igual a X.

cantO :: String -> Int

Cuenta la cantidad de O que tiene el estado de juego.

Suma 1 cada vez que un caracter de la cadena es igual a O.

cantE :: String -> Int

Cuenta la cantidad de E que tiene el estado de juego.

Suma 1 cada vez que un caracter de la cadena es igual a E.

ganador :: String -> Char

Verifica cada línea vertical, horizontal y diagonal en busca de un ganador.

Si lo encuentra devuelve el ganador (X u O), sino devuelve un caracter vacío.

Primero verifica que el comienzo de cada fila no sea una E (empty) y luego que los valores

de esa línea sean iguales. Si lo son devuelve ese valor que puede ser X o O.

De igual manera se procede con las columnas y con las diagonales.

proxJugador :: String -> Char

Define que jugador tiene el siguiente turno usando cantX y cantO

El jugador elegido para ser el siguiente será el que tenga un movimiento menos.

Nótese que en caso de igualdad siempre el jugador O mueve primero.

movPermitidos :: String -> [Int]

Arma una lista con todas posiciones vacías.

Utiliza la función ganador y la función esValido

Recibe una cadena con un estado de juego, si existe un ganador devuelve una lista vacía

que representa que no hay movimientos válidos.

De lo contrario usando la función esValido devuelve una lista de enteros en las

posiciones donde todavía puede jugarse.

esValido :: String -> Int -> Bool

Retorna verdadero si el movimiento es válido

Recibe un entero que verifica que esté entre 0 y 9, y un estado de juego.

Si el valor coincide donde se encuentra el caracter E devuelve verdadero, de lo contrario

devuelve falso.

mover :: String -> Char -> Int -> String

Esta función realiza el movimiento del jugador a la posición enviada y devuelve un nuevo tablero

actualizado donde se encuentra representado el último movimiento.

Recibe un estado del juego, X y/o O y una posición válida.

Del estado recibido separa el primer valor y si este es 0 (otherwise) se marca el primer lugar

con el caracter enviado (X,O).

Si el valor es mayor que 0 comienza una llamada recursiva. Empieza a llenar una lista vacía con el primer

valor del String en cada llamada recursiva, al tener un valor menos la posición se decrementa en 1 hasta

llegar a 0 donde se coloca el caracter recibido. Al final se concatenan los caracteres individuales extraídos

en cada llamada recursiva, el caracter enviado y el resto del String.

mejorMovimiento :: String -> Int

Retorna una posición que asegura no se pierde. Recibe un estado de juego y devuelve un entero.

Utiliza la función puntajeMovimientos que devuelve una lista de tuplas que contiene el movimiento y

su puntaje. Luego invoca a maximoPuntaje que devuelve el movimiento que posee mayor puntaje.

para obtener la posición deseada. La función foldr toma el primer valor de mov_puntaje y aplica la

función con el último elemento de la lista (tail mov_puntaje), luego aplica la función con el resultado

de la anterior y el penúltimo valor de la lista. Y así sucesivamente.

puntajeMovimientos :: String -> [(Int, Int)]

Retorna una lista de tuplas (movimiento, puntaje)

Invoca a movPermitidos y a mayorPuntaje.

Con la función zip se agrupan en tuplas cada movimiento con su respectivo puntaje.

en "tableros" (lista) se guardan tantos tableros como movimientos posibles haya. Cada uno

representando uno de esos movimientos.

En puntajes (lista) se guardan los puntajes de los estados de juegos guardados en la lista "tableros"

mayorPuntaje :: String -> Int

Retorna el mayor puntaje de movimiento para el tablero recibido.

Invoca a movPermitidos, puntajeJugador, mover y menorPuntaje

Sino existen más movimientos por hacer se fija si el jugador O es el ganador o no, y recibe un puntaje

de acuerdo al resultado.

Al igual que antes en "tableros" se guardan todos los posibles movimientos con una longitud igual a

la cantidad de E que haya.

En este punto las funciones mayorPuntaje y __menorPuntaje__ comienzan a llamarse estre sí

hasta cubrir con la totalidad de movimientos que quedan por delante hasta finalizar el juego,

intercalando entre los jugadores.

En otherwise se ordenan los puntajes, cuando se juega X se considera el menor puntaje, pues

es el que hace perder al jugador O, de lo contrario se considera el mayor.

menorPuntaje :: String -> Int

Retorna el menor puntaje de movimiento para el tablero recibido

Invoca a movPermitidos, puntajeJugador, mover y mayorPuntaje

Sino existen más movimientos por hacer se fija si el jugador O es el ganador o no.

Al igual que antes en "tableros" se guardan todos los posibles movimientos con una longitud igual a

la cantidad de E que haya.

puntajeJugador :: String -> Char -> Int

Asigna un puntaje a un jugador recibido.

Usando la función ganador, si no hay ganador devuelve 0, si el ganador es el jugador que se

recibió (Char) devuelve 10, sino lo es devuelve -10.

Como se vió antes, siempre se comprueba usando al jugador O.

maximoPuntaje :: Ord b => (a, b) -> (a, b) -> (a, b)

De dos tuplas (movimiento, puntaje) devuelve la de mayor puntaje.